requirements.txt
streamlit==1.31.0
PyMuPDF==1.23.26
sentence-transformers==2.3.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
openpyxl==3.1.2
pandas==2.2.0
python-dotenv==1.0.1
requests>=2.31.0
